requirements.txt
streamlit
yfinance
pandas
plotly
requests
python-dotenv
openai
